backend/requirements.txt
fastapi
uvicorn
motor
firebase-admin
pydantic-settings
python-multipart
requests
httpx
email-validator
python-dotenv
beautifulsoup4
deep-translator


